Move storyquestions assets in temporary folder
