+++ static/src/javascripts/bootstraps/enhanced/article.js
@@ -13,7 +13,7 @@ import {
 import { upgradeMembershipEvents } from 'common/modules/article/membership-events';
 import { geoMostPopular } from 'common/modules/onward/geo-most-popular';
 import { handleCompletion as handleQuizCompletion } from 'common/modules/atoms/quiz';
-import { init as initStoryQuestions } from 'common/modules/atoms/story-questions';
+import { init as initStoryQuestions } from 'journalism/temporary/story-questions';
 import { init as initLiveblogCommon } from 'bootstraps/enhanced/article-liveblog-common';
 import { initTrails } from 'bootstraps/enhanced/trail';
 import ophan from 'ophan/ng';
