+++ static/src/stylesheets/_common.scss
@@ -20,7 +20,7 @@
 @import 'module/content/_media-player';
 @import 'module/atoms/_youtube';
 @import 'module/atoms/_media';
-@import 'module/atoms/_storyquestions';
+@import 'module/atoms/temporary/_storyquestions';
 @import 'module/atoms/temporary/_explainers';
 @import 'module/atoms/temporary/_snippets';
 @import 'module/commercial/_commercial';
